refactor : 수동, 자동 로또 생성 책임 LottoFactory에 위임
